requirements.txt
tensorflow>=2.12.0
streamlit>=1.24.0
pillow>=9.5.0
numpy>=1.23.5
plotly>=5.14.1
pandas>=2.0.2
seaborn>=0.12.2
matplotlib>=3.7.1
scikit-learn>=1.2.2